Add files via upload
Sometimes, when selling something through OLX, you may become a victim of phishing. You need to be careful; this is what today's internet usage demands. However, you can also play a little trick on these damn thieves - flood their API.

How to do it?

When you receive a link with an address to a website resembling OLX or InPost, where you're asked to provide login details for your bank, go to the website in incognito mode with the browser's traffic monitor enabled. Choose any bank and fill in the login details (use fake data!!!!). Extract the POST request from the browser and paste it into Postman as cURL. In the pre-request script, I added a script that generates a new random transaction number with each request sent, consistent with what the thieves were generating. Additionally, you can add scripts here to generate random logins and passwords, but in my opinion, flooding with the number of POST requests with unique transaction IDs will do the trick.

How to use it?

Just run the entire collection in Postman and execute, for example, 9999 iterations ;-) Have fun!